requirements.txt
absl-py==2.4.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.37
etils==1.13.0
filelock==3.24.2
fsspec==2026.2.0
glfw==2.10.0
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
labmaze==1.0.6
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
mujoco==3.5.0
networkx==3.6.1
numpy==2.4.2
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079150/work
pillow==12.1.1
protobuf==6.33.5
PyOpenGL==3.1.10
pyparsing==3.3.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
setuptools==80.10.2
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
wrapt==2.1.1
zipp==3.23.0
